requirements.txt
absl-py==2.1.0
ai-edge-litert==1.0.1
apache-beam==2.62.0
array_record==0.5.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
avro-python3==1.10.2
bleach==6.2.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.0
crcmod==1.7
cycler==0.12.1
Cython==3.0.11
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
dm-tree==0.1.8
dnspython==2.7.0
docopt==0.6.2
etils==1.5.2
exceptiongroup==1.2.2
executing==2.1.0
fastavro==1.10.0
fasteners==0.19
filelock==3.17.0
filetype==1.2.0
flatbuffers==25.1.24
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
gast==0.4.0
gin-config==0.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.65.5
h5py==3.12.1
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
ipykernel==6.29.5
ipython==8.18.1
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpickle==3.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.6.17
keras==2.15.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
labelImg==1.8.6
libclang==18.1.1
lvis==0.5.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauth2client==4.1.3
oauthlib==3.2.2
object-detection @ file:///home/nitin/TF%20Model%20Garden/models/research
objsize==0.7.1
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.13.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.1.1
promise==2.3
prompt_toolkit==3.0.48
proto-plus==1.26.0
protobuf==3.20.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pydot==1.4.2
Pygments==2.18.0
pymongo==4.10.1
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
roboflow==1.1.53
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
seqeval==1.2.2
six==1.15.0
sortedcontainers==2.4.0
stack-data==0.6.3
supervision==0.25.1
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow @ https://files.pythonhosted.org/packages/dc/cf/e7e9888e3a22fe1fcdef23a811bd7d47504eb1e128a3076e1d2523ddb0ab/tensorflow-2.15.0-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl#sha256=3fa865956d96b7614f247c36e4c22b1543ba5ce656fbe8e4f6266ae7a4917132
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-model-optimization==0.8.0
tensorflow-text==2.15.0
termcolor==1.1.0
text-unidecode==1.3
tf-models-official==2.15.0
tf-slim==1.1.0
tf_keras==2.15.1
threadpoolctl==3.5.0
toml==0.10.2
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.40
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.2.3
virtualenv-clone==0.5.7
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.12.1
zipp==3.21.0
zstandard==0.23.0
